+++ sourcespec/ssp_inversion.py
@@ -308,14 +308,17 @@ def _spec_inversion(config, spec, spec_weight):
         confidence_level=68.2, format='{:.3f}')
 
     # additional parameters, computed from fc, Mw and t_star
-    vs = config.event.hypocenter.vs
+    if config.wave_type == 'P':
+        vel = config.event.hypocenter.vp
+    else:
+        vel = config.event.hypocenter.vs
     travel_time = spec.stats.travel_times[config.wave_type[0]]
     # seismic moment
     station_pars.Mo = SpectralParameter(
         id='Mw', value=mag_to_moment(Mw), format='{:.3e}')
     # source radius in meters
     station_pars.radius = SpectralParameter(
-        id='radius', value=source_radius(fc, vs * 1e3), format='{:.3f}')
+        id='radius', value=source_radius(fc, vel * 1e3), format='{:.3f}')
     # Brune static stress drop in MPa
     station_pars.bsd = SpectralParameter(
         id='bsd', value=bsd(station_pars.Mo.value, station_pars.radius.value),
@@ -347,8 +350,8 @@ def _spec_inversion(config, spec, spec_weight):
     if fc_min <= 0:
         fc_min = freq_log[0]
     fc_max = fc + fc_err[1]
-    radius_min = source_radius(fc_max, vs * 1e3)
-    radius_max = source_radius(fc_min, vs * 1e3)
+    radius_min = source_radius(fc_max, vel * 1e3)
+    radius_max = source_radius(fc_min, vel * 1e3)
     station_pars.radius.lower_uncertainty =\
         station_pars.radius.value - radius_min
     station_pars.radius.upper_uncertainty =\